Team -> Controller -> CRUD básico ok

+++ src/main/java/com/svalero/happDeporte/Util/Constants.java
@@ -2,7 +2,7 @@ package com.svalero.happDeporte.Util;
 
 public class Constants {
 
-    public static final float QUOTA = 300f;
+    public static final double QUOTA = 300;
     public static final float PRICE_EQUIPMENT = 55.50f;
     public static final float PRICE_SWEATSHIRT = 25f;
 }
